[Lodgify.Cinema.Domain] Namespace refactor
