+++ Lodgify.Cinema.Domain/Entitie/ShowtimeEntity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace ApiApplication.Database.Entities
+namespace Lodgify.Cinema.Domain.Entitie
 {
     public class ShowtimeEntity
     {
